uhd: Change Unknown PPS behavior
Instead of setting USRP time to 0 with `time_unknown_pps`, set it to current
time. This does not help with fine sync but it helps to stay in the same
ball park with all clocks.
One might accidently send timed commands with PC clock time. A USRP
would try to wait 50 years at this point to actually execute that
command or to send that burst if the time at given in case of "Unknown
PPS". This may cause issues during shutdown because it may force a user
to send a kill signal to the flowgraph. In turn that may require a USRP
reboot.