digital/symbol_sync: Remove manual memory management
The `cc` version already used more smart pointers, but `d_clock`
didn't need to be a pointer at all.

I consted function args to make it clear in the initializer that they
will not be changed in the constructor body.

Return value for `make` for the TED and interpolator was changed to a
`unique_ptr` to make it clear that ownership is transferred. And it
makes it all but impossible to accidentally memory leak.
(core guidelines F.26. Also relevant I.11,R.20).