vocoder: refactor get_next_tx_char callback
The get_next_tx_char callback in the FreeDV modulator block is one of 
the last places where snprintf is used. I've eliminated it by using C++ 
string methods instead. This also removes the 79-character limit.

This change also fixes a bug where the last character of the message was 
not transmitted.

Lastly, I refactored the callback to be a static member function, 
eliminating the need for a separate structure to hold the callback 
state.