grc: remove gnome-terminal deprecation warning
When running a "No GUI" flowgraph from GRC, it gives the following
deprecation warning:
Option “-e” is deprecated and might be removed in a later version of gnome-terminal.
Use “-- ” to terminate the options and put the command line to execute after it.

This follows the suggested fix to update the deprecation

Signed-off-by: Josh Morman <jmorman@perspectalabs.com>