filter: Fix polyphase_filterbank for MSVC, w/ default move constructor.
Because of the deleted copy constructor (previously implicit, now made
explicit in this commit to match fir_filter.h), polyphase_filterbank.cc
fails to compile with MSVC trying to use the fft_filter_ccf copy
constructor. Explicitly declaring a default move constructor fixes this
by allowing it to use the move constuctor instead, which is the
behavior we want. It's a mystery why MSVC requires this and other
compilers don't, but it works.

See #3892 for similar discussion and change for other filter classes.

Signed-off-by: Ryan Volz <ryan.volz@gmail.com>