gr-digital: Fixup GUI hints for example_corr_est
The tab control were missing a proper GUI hint and was just placed in
the bottom corner.

Add GUI hint to it and adjust the amount of rows the other time sink
consumes to let the sliders not consume that much screen real estate.

Slightly align blocks to look nicer.

Signed-off-by: Nick Østergaard <oe.nick@gmail.com>