Pybind: Adding header hash hotfixer
DO NOT USE THIS TOOL IF ADDING OR REMOVING ANY FUNCTIONALITY FROM A
HEADER.

This tool allows to quickly update the hash in a binding file
(*/python/*/binding/something.cc).

WARNING: This script updates header hashes without checking whether the
underlying functionality changed. If you added or removed any python-bound
function in the header to this binding file THIS IS NOT THE TOOL YOU SHOULD BE
USING. Use bindtool instead to generate correct binding code.

Signed-off-by: Marcus Müller <mmueller@gnuradio.org>