gr-filter: update for additional window values
Recently added window options include several additional windows that take a configuration parameter. This updates this parameter from 'beta' which was previously used only for Kaiser windows to a generic 'param' which makes sense for any window requiring a configuration parameter. The default values remain 6.76 which is not legal for the Tukey window but this error will be caught if the default is used; this is a reasonable cost for backward compatability.

Signed-off-by: Jacob Gilbert <jacob.gilbert@protonmail.com>