cmake: Format numpy include path in CMake style for consistency.
GNU Radio builds done with conda do path subsitution to replace the
build prefix with the installed prefix, but it struggles when multiple
path separator styles are used in the same file. This makes it so that
all of the paths that would be subsitituted use the same style (forward
slash separators) so that build prefix replacement happens correctly.

Signed-off-by: Ryan Volz <ryan.volz@gmail.com>
(cherry picked from commit 8793effe5d6013c95b44f52551ba0d660157b263)
Signed-off-by: Jeff Long <willcode4@gmail.com>