grc: Be more informative when missing "options" block.
This raises an RuntimeError describing that GRC could not find the
built-in blocks rather than raising a KeyError on "options", which
left users with little to go on. This error has come up in various
situations involving an incorrect blocks path (due to mixing GR
versions causing the configuration files to point to the wrong
installation) or simply the blocks not being installed.

Signed-off-by: Ryan Volz <ryan.volz@gmail.com>
Signed-off-by: Bernard Tyers - Sane UX Design <bernard+work@ei8fdb.org>