grc: Remove global_blocks_path preference and use prefix path.
Instead of storing the default (built-in) blocks path in the
preferences, store the path relative to the prefix (GRC_BLOCKS_DIR) in
constants.yml file and configure it at runtime using the installation
prefix. This should make it easier to switch between different versions
by making it so the user preference for global_blocks_path doesn't
interfere. Users still have local_blocks_path preference and the
GRC_BLOCKS_PATH environment variable for pointing to additional blocks.

Signed-off-by: Ryan Volz <ryan.volz@gmail.com>