gr-digital: Replace CMA Equalizer in MPSK Example
In commit 65bd288, the `digital_cma_equalizer_cc` block was removed.
However, it is still left in the MPSK example flowgraph. This commit
replaces the deprecated cma equalizer with a linear equalizer and an
adaptive algorithm as those are indicated in the gnuradio wiki to be the
appropriate replacements. I have kept the parameters between the two
sets as similar as possible by ensuring the parameters with the same ids
have the same values. The only difference is `Modulus`.
Amended `Modulus` from 1 to 4. A note was also added to indicate the
change from the CMA Equalizer to the Linear Equalizer.

Signed-off-by: Solomon Tan <solomonbstoner@yahoo.com.au>
(cherry picked from commit 5cdb832ed4ca7840e1d1c9efc70e55d2647654c4)
Signed-off-by: Jeff Long <willcode4@gmail.com>