grc: handle variables of type raw correctly
At the moment grc does not validate variables, as they are of type raw.
So they are always treated as valid, which leads to some problems, see: #4519.

A approach to fix #4519 is #4531.
This approach improves the situation, but has still some problems.

param.py provides a validator function for dtype raw, but this is not used, as dtypes.py provides no validator function for dtype raw.

This fix clears the error messages in base.rewrite not in base.elements.validate, so the error messages are kept.

In addition this makes it possible, to make the code in VariableEditor.py smoother.

The error messages are kept on switching the different tabs in VariableEditor.

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>
(cherry picked from commit e562d6b3c82a30f12b4e8e2585447dddeb88e93d)
Signed-off-by: Jeff Long <willcode4@gmail.com>