digital: Remove crc32 utility
This patch removes a crc32 utility that used to be used in our CRC32
blocks, but no longer is. Right now, all this does is expose a generic
CRC32 function to the user (both via Python and C++), but there are
enough other libraries that let the user do that. Therefore, we'll
remove this.

Signed-off-by: Martin Braun <martin@gnuradio.org>
Signed-off-by: Bernard Tyers - Sane UX Design <bernard+work@ei8fdb.org>