modtool: Fix Issue #2672 for `disable` & more
When the user specifies a block name, only the block with that specific
blockname should be disabled. Prior to this PR, all blocks matching said
specified block name are disabled. This PR fixes it.

This PR also fixes what appears to be a spelling error `self._info`. The
error is thrown for OOT module version 38:
Attribute error `'ModToolDisable' object has no attribute '_info'`

Signed-off-by: Solomon Tan <solomonbstoner@yahoo.com.au>