grc: add core features of view-only mode
These additions introduces the backend parts of the View-Only Mode
which is a new security feature of the GRC. It disables untrusted
evaluation in flowgraphs, such as parameter evaluation. It
exports saved values of parameters and template expressions, saving
them directly in the .grc-file and having untrusted flowgraphs use
those saved values instead of performing evaluation.

Flowgraphs from files are untrusted whereas new flowgraphs are trusted,
but only within that instance of the GRC as trust is currently not
persistent.

This introduces a custom exception, a TrustError, which is used
whenever the View-Only Mode fails, for example if a file lacks the
saved values needed by the View-Only Mode. This could later be used
to trigger a prompt which asks the user to trust the flowgraph.

This commit makes up most of the backend parts of the mode, which
will later be accompanied by GUI elements and additional features.

Signed-off-by: oscar ekholm <oscarekh@kth.se>