gr-newmod: Set OOT module VERSION_PATCH to 0 instead of git.
This avoids creating multiple copies of the .so file
(each with a different git tag) during development.
The ldconfig command doesn't necessarily set the symlink
properly and the old .so can get loaded after an update.
This can be very confusing for developers and this patch
suppresses that behavior by always creating the .so with
the same name.

Signed-off-by: Ron Economos <w6rz@comcast.net>