grc: Allow func prob args to be changed dynamically
When the func prob block is used to run a function defined in a epy
block to accept any number of arguments, it is impossible to dynamically
change the value of the argument during runtime despite a callback
function setting its value. This is because it is missing a 'self.'
prefix.

This commit ensures each argument provided is prepended with a 'self'.
An example will be provided with the PR to illustrate the changes.

Signed-off-by: Solomon Tan <solomonbstoner@yahoo.com.au>