filter: use output multiple for pfb arbitrary resampler
When interpolating, the resampler was taking up a full cpu even
at very low rates. This is due to polling when no input items are
available.

For interpolation, this change sets the output multiple to the greater
of the relative rate and the number of pfb filter arms. This is fairly
arbitrary, but results in usable output multiples and good performance.

The output multiple can only be set effectively in the block constructor.
This means that it will not change if the relative rate is set
dynamically. For widely varying rates, this could be a problem. For
rates that vary around 1.0, if the rate is initially set to 0.9 and
then changed to 1.1, performance will be as before (bad) when the rate
goes above 1.0. For this reason, and in hopes that the user will start
with 1.0 in such a case, the output multiple is set for 1.0.

Signed-off-by: Jeff Long <willcode4@gmail.com>