runtime: deprecate tag_checker (removal in 3.10)
tag_checker used to be meant for cases where you've gotten an unsorted
tag vector (mostly, from get_tags_in_range), which you then had to go
through in parallel to your samples, to check which tag applies at what
sample.

For that it sorts the tags; the tags coming from the get_tags* functions
are sorted already

The checking pattern (which chunks_to_symbols is the last consumer of)
is inefficient: instead of taking the index of the first unprocessed tag
and processing all samples up to that index, a check is performed on
every sample, which includes calls and multiple indirections.

So, since very likely nobody uses this, and because it's a design
anti-pattern, deprecating this.

Signed-off-by: Marcus Müller <mmueller@gnuradio.org>