runtime: logging performance improvement
Based on email discussion with Boris Marjanovic.

The logging macros always pass the "msg" param to the logger (currently
log4cpp). In the macros, "msg" can be an expression and it will be
evaluated, even if the log message would not be printed (based on log
level).

This commit adds checks to each macro to make sure the message will be
logged. Since this check is also done internally to log4cpp, some
duplication results, but it should be far less effort than evaluating
an arbitrary "msg".

Signed-off-by: Jeff Long <willcode4@gmail.com>