grc: Generate c++ vector from list compre in var
Issue 4955 states that when a list comprehension is used in a variable
block, c++ code cannot be generated from the flowgraph. The cause of the
problem is `ast.literal_eval` which only considers a small number of
Python syntaxes to be valid. asteval accepts a greater range of syntaxes
while being safer than `eval`.

Signed-off-by: Solomon Tan <solomonbstoner@yahoo.com.au>