cmake: Fix cmake failure with unset VERSION_PATCH.
After the recent change to use `string(REGEX REPLACE ...)` to ensure
that only numeric values are used for RC_*_VERSION, an unset
VERSION_PATCH now results in a cmake failure when it didn't previously.
While we should expect all 4 version fields to be set (except when one
of the values is "git"), we should also avoid failing here when it
worked before. At least one OOT (gr-satellites) doesn't set the
VERSION_PATCH field for releases, so this fixes the build in that case
without requiring OOT changes. Now an unset VERSION_PATCH defaults to
"0".

Signed-off-by: Ryan Volz <ryan.volz@gmail.com>