iio: GRC bindings: remove global variable from hidden defaults
When the fields are hidden, and the expected global variable is
undefined, the flow graph becomes unsynthesizable.

I've only found this pattern with IIO blocks.

Sadly, we *have* to put defaults here, in order to have valid arguments
to the (useless) set_filter_params call in the non-Design filter mode.

Since the fpass and fstop fields, however, aren't used, the actual value
used there doesn't matter – so, 0 it is.

Signed-off-by: Marcus Müller <mmueller@gnuradio.org>