gr-qtgui: Frequency sink freezes the running flowgraph
With #5405 the segfault of freq_sink_c was fixed, but
if you enable the control panel and change the FFT size to 8192
while the flow graph is running, the whole flowgraph freezes.

This patch:

- fixes the freeze
- extends the fftsize to 32768 for freq_sink_c and freq_sink_f
- changes the entry for fftsize in the corresponding yml files from integer to enum
  and provides a drop down menu with those values that are provided by the gui
  in the running flowgraph.

The code provides code for setting the fftsize in form_menus.h and in freqcontrolpanel.cc
If you change some signal and slots you can do without the code from form_menus.h.
So this part is removed.

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>