gr-qtgui: Frequency sink freezes the running flowgraph
* gr-qtgui: Frequency sink freezes the running flowgraph

With #5405 the segfault of freq_sink_c was fixed, but
if you enable the control panel and change the FFT size to 8192
while the flow graph is running, the whole flowgraph freezes.

This patch:

- fixes the freeze
- extends the fftsize to 32768 for freq_sink_c and freq_sink_f
- changes the entry for fftsize in the corresponding yml files from integer to enum
  and provides a drop down menu with those values that are provided by the gui
  in the running flowgraph.

The code provides code for setting the fftsize in form_menus.h and in freqcontrolpanel.cc
If you change some signal and slots you can do without the code from form_menus.h.
So this part is removed.

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>

* Keep fftsize list, but move type back to int

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>

* Use const vars for fft range.
Build the fftsize list from those vars

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>

* As the fftsize must not be taken from the dropdown menu
but can be set to a variable, check if the value belongs
to the dropdown menu, otherwise select the greatest value
in the menu < given fftsize.

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>

* qtgui: fixed error text on freq sinks

Signed-off-by: Josh Morman <jmorman@gnuradio.org>

* qtgui: add assert into freq grc file

Co-authored-by: Josh Morman <jmorman@gnuradio.org>
(cherry picked from commit 3452e657c0c96db66770cc4e06045e5c704a0242)
Signed-off-by: Jeff Long <willcode4@gmail.com>