grc, block bindings, enable additional error messages for assertions
If the validation of an assert statement of a block fails,
grc only shows the failed assertion. This may sometimes be a bit cryptic like

Assertion "not (fftsize & (fftsize -1))" failed

which means  "fftsize must be a power of 2".

So it would be nice, to add optional a descriptive message to the assertion.

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>
(cherry picked from commit 0095840b10d6a5894286e2b3120acd7ed7d46981)
Signed-off-by: Jeff Long <willcode4@gmail.com>