Fix iio sink blocks packet_len processor
Packet length tagged processing would break for FMComms234 sinks due to
how the work function gets called and channels are interpreted.
Borrowing from the fix for the source blocks an additional property is
used to tell the parent device_sink block what the true channel inputs
are to be processed.

Signed-off-by: Travis F. Collins <travis.collins@analog.com>