cmake: Initialize `CMAKE_BUILD_TYPE` to Release before project() call.
Calling project() sets the default value for `CMAKE_BUILD_TYPE`. For
most generators this results in a null value so that any later check
acting on an unset `CMAKE_BUILD_TYPE` will proceed. However, on Windows
the default value is set to Debug, so the existing attempts to detect an
unset `CMAKE_BUILD_TYPE` and set it to Release fail to do the job on
Windows.

This commit sets `CMAKE_BUILD_TYPE` to Release before the project() call
so that the generator-dependent default is avoided. Because
`CMAKE_BUILD_TYPE` is set as a `CACHE` variable, this set command will
not overwrite any existing cached value for `CMAKE_BUILD_TYPE`, thus
still allowing users to override the setting from the CMake
command line.

Signed-off-by: Ryan Volz <ryan.volz@gmail.com>
(cherry picked from commit fa67b36027de1368d0e65760148f50b9786459d5)
Signed-off-by: Jeff Long <willcode4@gmail.com>