grc: store grc version info in metadata
Saving a flowgraph does not save the grc version
as it this info is not part of the schema.

This pr adds grc_version to the schema definition and
saves the grc version string when the fg is saved.

Possibly
fixes #5521

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>
(cherry picked from commit b0d3304e1b82821b60027858eb220bc513e2cf6f)
Signed-off-by: Jeff Long <willcode4@gmail.com>