cmake: Remove 'REQUIRED' flag for Volk
This makes VOLK an official dependency of gnuradio-runtime, but no
longer fails to execute CMake if VOLK is not found. This enables
standalone building of targets such as the manual, which do not depend
on the runtime.

Signed-off-by: Martin Braun <martin@gnuradio.org>
(cherry picked from commit 776a0c35c66004e12f15301e6ad52bc8292e2567)
Signed-off-by: Jeff Long <willcode4@gmail.com>