digital: add generic CRC blocks
This adds a crc class that implements a generic CRC calculation using
a table-driven byte-by-byte approach. The CRC parameters are set when
instancing the class.

The crc class is used in the CRC Append and CRC Check blocks, which
take a PDU as input, and either compute the CRC of that PDU and append
it, or check the CRC included in the PDU (sending out the PDU in one
of two outputs according as to whether the CRC was correct).

The CRC blocks can optionally ignore a certain amount of "header" bytes
in the CRC calculation (used for packets where the CRC only affects the
"payload" and not the header). Additionally, the CRC Check block can
optionally remove the CRC after checking it.

Signed-off-by: Daniel Estévez <daniel@destevez.net>
(cherry picked from commit 1c4ef3dac7a12d4841b9544cc4c6cef0aca4cff0)
Signed-off-by: Jeff Long <willcode4@gmail.com>