runtime: Make vmcircbuf logging quieter
When a flow graph is executed for the first time, and ~/.gnuradio does
not exist, the following error messages are printed:

vmcircbuf_prefs::get :erro: /home/argilo/.gnuradio/prefs/vmcircbuf_default_factory: No such file or directory
gr::vmcircbuf :error: vmcircbuf_createfilemapping: createfilemapping is not available

These are expected situations, so they should not be logged at the error
level. Other similar vmcircbuf messages are logged to the debug logger
at the info level, so I've done that here as well.

In addition, I changed createfilemapping to CreateFileMapping to reflect
the actual name of the function.

After these changes, no log messages are printed the first time GNU Radio
is used, since the debug logger is set to the emerg level by default.

Signed-off-by: Clayton Smith <argilo@gmail.com>