cmake: Always prefix git hash used as version with "g".
This prevents the hash from being misinterpreted as a number, which MSVC
does which will caused failed builds. This is never an issue with GR
itself because the plain hash is only used for the version when the
repository is untagged, but it can happen with OOTs which also use
GrVersion.cmake.

This works by removing the `--always` flag from `git describe` and
handling the no-tag case as a separate special case.

I have been unlucky enough to twice now encounter all-numeric short
hashes when building untagged OOTs with MSVC, so I can't ignore it.

Signed-off-by: Ryan Volz <ryan.volz@gmail.com>
(cherry picked from commit fa8953de2ae67990ccb052af08affac073e106ae)
Signed-off-by: Jeff Long <willcode4@gmail.com>