filter: Fix flaky qa_filterbank test
The qa_filterbank test occasionally fails, and from the error message
it's apparent that on these occasions outdata2 contains the same values
as outdata. This suggests that the while loop that waits for the new
filter taps to kick in is exiting too soon. I suspect this occurs
because it compares floating point values for exact equality. I've
changed this to check for approximate equality instead.

Signed-off-by: Clayton Smith <argilo@gmail.com>