blocks: Add symbolinterleaver block
This commit adds a symbol interleaver to gr-blocks.
The idea is to specify an interleaver sequence to be able to interleave
and deinterleave vectors of constant size. For convenience, the
deinterleaver sequence is always derived. Thus, it is not necessary to
manually derive the corresponding deinterleaver sequence.

See issue #3637 for a preliminary discussion.
https://github.com/gnuradio/gnuradio/issues/3637

Signed-off-by: Johannes Demel <demel@ant.uni-bremen.de>