Fix various broken fmt strings
In a few instances, the format string does not match the types of the
arguments. In other cases, the format string is not constant. I fixed
both problems here, and verified that C++20 type checking is happy
with the updated format strings.

Signed-off-by: Clayton Smith <argilo@gmail.com>
(cherry picked from commit bed4dba514bb9dbd48defbb744debb77c6710202)
Signed-off-by: Jeff Long <willcode4@gmail.com>