blocks: Fix rotator_cc scheduled phase inc updates
The previous implementation was applying the rotation to the wrong
indexes when multiple scheduled phase increment updates were processed
in the same work call. This patch fixes the input/output buffer indexes
on the `rotateN()` call and covers the multi-update scenario on QA. The
previous QA implementation was not covering this bug because the
multi-update test case was only checking the tags produced on each phase
increment update and not whether the actual IQ output was correct.

Signed-off-by: Igor Freire <igor@blockstream.com>
(cherry picked from commit c1d882f49ab54666f5e232bbef670d729587a84b)
Signed-off-by: Jeff Long <willcode4@gmail.com>