cmake: Change iio and ad9361 targets to interface to fix macOS linking.
Previously, if either libiio or libad9361 were found as a framework
bundle on macOS, linking would fail because it would try to link to the
framework directory and not the actual library file. This change sets
the found library as an interface library on the imported target, which
lets CMake automatically handle the case of a framework versus a normal
library file.

Signed-off-by: Ryan Volz <ryan.volz@gmail.com>
(cherry picked from commit b4931e1ec3410a89cd89a15e0542d6d2b2f835ba)
Signed-off-by: Jeff Long <willcode4@gmail.com>