cmake: make building with clang possible again
This required quite a bit of experimentation. When using clang on a
Linux (i.e., GCC-dominated) system, it uses the libgcc_s C++ runtime.
Which doesn't contain atomics, and that hurts at later times when
linking GCC-built libraries that assume you'd be -latomic anyways
(looking at Boost *again*).

The sad part is that we need some kind of unpredictable mixed
compilation to trigger this. So a CMake-time compile check doesn't work.

To the best of my knowledge, this works on newer and older clangs on
arbitrary platforms. I'm still checking for Linux here; if the same
problem happens on some xBSD or other platform, it should be easy
enough to find the solution for

undefined reference to `__atomic_store'

being exactly this Linux hack.

Signed-off-by: Marcus Müller <mmueller@gnuradio.org>