4.0: Make gnuradio installation path consistent
On gnuradio 3.x there is only on 'prefix'/lib(64) with  its subdirectories.
On 4.0 you can set --libdir. Setting this to lib64, you get
'prefix'/lib64 and
'prefix'/lib
 with subdirectories.

This patch removes the hardcoded lib subdirectory path and forces to install only to one directory ( lib or lib64 ) depending on the --libdir setting.

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>