uhd: rfnoc: Update block controllers to match gr-ettus
This puts gr-uhd at feature-parity with gr-ettus for the following
blocks: FIR filter, siggen, and window. It also updates an issue in
rfnoc_block regarding setting of properties.

The FIR filter has different features starting in UHD 4.2. This is
checked with a compiler flag.

GRC Examples for those new blocks are also updated.

Signed-off-by: Martin Braun <martin.braun@ettus.com>