blocks: Fix file tests to work on Windows
As noted in the documentation for NamedTemporaryFile, it is not possible
on all platforms to open the temporary file a second time before closing
it. Thus, we need to be sure to close the file before opening it in a
flowgraph.

Signed-off-by: Clayton Smith <argilo@gmail.com>
(cherry picked from commit 1142d1679ce8e199451207444d6bc2336b07286d)
Signed-off-by: Jeff Long <willcode4@gmail.com>