runtime: Move block_gateway from gnuradio-runtime into gr_python
This eliminates the need to link gnuradio-runtime with Python and fixes
segfaults on conda+macOS when trying to use any Python block. Those
segfaults would happen because the Python library was loaded twice: once
dynamically from linking with gnuradio-runtime, and once from the
running Python interpreter which conda-forge links statically with
libpython. The block_gateway was only used from its Python binding, so
the implementation now just lives there.

This is an API change!

Signed-off-by: Ryan Volz <ryan.volz@gmail.com>