uhd: rfnoc: Generate correct Python when using clock/time source
In GRC, there was an issue where you could select a clock/time source
for the RFNoC graph, but it was never passed on to the graph object.
This fixes the issue such that the correct device args are generated.

Signed-off-by: Martin Braun <martin.braun@ettus.com>
(cherry picked from commit 86b5ab2898b3284345cf936eb519cc9fcc22da8e)
Signed-off-by: Jeff Long <willcode4@gmail.com>