grc: Don't query and use display scale factor in DrawingArea size.
All this did was request a larger size for the cairo surface of the
DrawingArea on which this is set, which very roughly worked around the
cairo context reference bug for users with high DPI displays. Now that
this bug is fixed by the previous commit, this is not necessary.

It's important to note that GTK takes care of the DPI scaling for high
DPI displays behind the scenes, so this doesn't even help to scale the
flowgraph properly. If there are further bugs with scaling, don't look
here.

Signed-off-by: Ryan Volz <ryan.volz@gmail.com>
(cherry picked from commit 28bbc390e490e087b32edee8ed072e469d345079)
Signed-off-by: Jeff Long <willcode4@gmail.com>