Allow freq message on gr-soapy source block cmd port to add rx_freq tag.
https://github.com/gnuradio/gnuradio/commit/3eea1057c53842476f809a62499ee6b8dc080d82
causes the gr-soapy block to add the rx_freq tag when set_frequency() is called with a name parameter.

However, gr-soapy source's cmd port handler calls set_frequency() without name, so add_tag_ is not
set and the rx_freq tag is not added. This patch causes rx_freq to be added for both versions of set_frequency().

Signed-off-by: Josh Bailey <josh@vandervecken.com>