filter: fix reversed arguments in vector initializer
This bug was creating a zero-length vector which was being passed to
fir_filter_with_buffer_ccf. A zero-length vector makes no sense for
that block, and on systems with certain debug settings an out-of-range
copy (of zero length) was detected, causing a crash.

There were also some reversed params in pfb_arb_resampler.cc.

Signed-off-by: Jeff Long <willcode4@gmail.com>