fec: fix regression in async_decoder and add tests
The async_decoder was broken in 4e50c99ab38f816d56145133ce6b7cc68e164e90.
The reason was a mistake when rewriting the PR to address comments.
In the case where the decoder input needs to be converted to uchar,
'nitems_out' rather than 'nbits_in' was wrongly used as the length
of the data to be converted. Effectively, we are passing only the
first half of the input to a rate=1/2 decoder such as the CC decoder.

Since there are no tests of the async_decoder that would have caught
the bug, this adds some tests. A new _qa_helper_async top_block similar
to _qa_helper is added. This is used in an additional test case in the
qa_fecapi_* tests to run the async_encoder and async_decoder with all the
different FECAPI encoders/decoders.

Signed-off-by: Daniel Estévez <daniel@destevez.net>
(cherry picked from commit 3c6717b85b42544aa95b883467927398e95f47bb)
Signed-off-by: Jeff Long <willcode4@gmail.com>