uhd: rfnoc: fix literal_eval error
literal_eval() was throwing an exception on drag and drop because it requires a string containing a literal (e.g. string) rather than just a string itself, it then returns the contained string.

This was only an issue on drag/drop as the block code later does this transformation itself, but not before the first var_make()

Signed-off-by: Grant Meyerhoff <grant.meyerhoff@ni.com>