uhd: rfnoc: fix literal_eval error
literal_eval() was used in the RFNoC graph GRC binding to sanitize
inputs. However, it was throwing an exception on drag and drop because
it requires a string containing a literal (e.g. string) rather than
just a string itself,

This was only an issue on drag/drop as the block code later does this transformation itself, but not before the first var_make()

However, the literal_eval() is not actually necessary, since the objects
passed in are of type TemplateArg, not just
string-or-something-that-evaluates-to-string, and thus we can use
facilities from TemplateArg to properly escape the values before usage.

Signed-off-by: Grant Meyerhoff <grant.meyerhoff@ni.com>
Co-authored-by: Martin Braun <martin.braun@ettus.com>
Signed-off-by: Martin Braun <martin.braun@ettus.com>