blocks: Add lambda_block
This is the C++ equivalent of the embedded Python block. It allows
passing in the code to general_work() as a parameter.

This block can be particularly useful when designing hierarchical blocks
which have a small custom kernel. Previously, this would require first
adding this custom kernel as a custom block, and then building the
hierarchical block. Now, custom work functions can be included in the
hierarchical block itself.

Co-authored-by: Grant Meyerhoff <grant.meyerhoff@ni.com>
Signed-off-by: Martin Braun <martin.braun@ettus.com>
Signed-off-by: Grant Meyerhoff <grant.meyerhoff@ni.com>