qtgui: Replace QLineEdit::returnPressed() with editingFinished()
This affects the QT GUI Entry, Message Edit Box, and Range widgets which
use QLineEdit for text entry. Previously, these text edit boxes would
not trigger the callback to update the variable unless 'Return'/'Enter'
was pressed. With this change, the variable is updated both in that case
and also when the field loses focus, e.g. by clicking somewhere else on
the display. See
https://doc.qt.io/qtforpython-5/PySide2/QtWidgets/QLineEdit.html#detailed-description
for a description of the two signals.

Signed-off-by: Ryan Volz <ryan.volz@gmail.com>