soapy: sdrplay agc and gain fixes
SDRPlay gains are actually attenuations
- grdb: gain reduction of 20 to 59 dB
- lna state: model-specific attenuation steps for LNA

LNA state is independent of AGC. Gain reduction is N/A
when AGC is on.

SoapySDR attempts to do automatic distributions of gains
if setGain() is called without a gain name. This is what
gr-soapy was trying to do and it resulted in invalid gain
values.

AGC has a setpoint, which was not being set. This is probably
why a crash was happening when AGC was set.

This change:
- Adds the AGC setpoint
- Separates gain and lna state
- Makes gain positive instead of the previous negative range
- LNA state is still an attenuation with device-specific stages
- Sets gain when agc is turned off (not done by driver)
- Defers setting IFGR gain when agc is on, to avoid warnings
- Enables bias tee control
- gain, lna_state, agc, agc_setpoint and bias are runtime settable

Signed-off-by: Jeff Long <willcode4@gmail.com>
(cherry picked from commit 8e2b21ba8f2d568be1e3c254b16020af39dd0142)
Signed-off-by: Jeff Long <willcode4@gmail.com>