blocks: Add a GRC example for throttle
Also used on the Throttle wiki page:

----

In the below flowgraph, the upper Signal Source is feeding into a Throttle block, which limits the rate at which it passes samples to 50,000 samples per second.
We use the "Limit" parameter to bound the maximum time that happens between a chunk of items being emitted by the Throttle block. We set that time to 1/50 of a second. (And that gives us 1/50 s · 50,000 Samples/s = 1000 Samples per chunk, at most.)
The lower Signal Source is not feeding into something that slows it down to a defined rate.
To test the rates at which samples pass through, we attach Probe Rate blocks, and give them distinct names. We use Message Debug to print these rates.

We observe this:

- As we can see in the upper "Throttled" figure, the passing of samples of a defined rate leads to predictable displaying of the waveform. For the unthrottled data, not so much
- We observe the sample "passing through" rates before the Throttle: It is, on average, as expected, 50,000 per second.
- We observe the rate between the unthrottled Signal Source and the Sink: that is in the region of 288 million samples a second. It's purely random which part of the waveform we see.
- The upper Signal source consumes negligible amounts of CPU. The lower one fully consumes a CPU core, and hence leads to heat and noise from CPU coolers.
- It is especially noteworthy that throttling in the top part of the flow graph has no effect on the bottom part.

Signed-off-by: Marcus Müller <mmueller@gnuradio.org>
(cherry picked from commit 0e69e7da7045d7a429c702bcb869a68b5ce2337b)
Signed-off-by: Jeff Long <willcode4@gmail.com>