blocks: repeat without output multiple constraints
Repeat is a block that's commonly runtime-reconfigured. However, if the new interpolation is larger than the one specified at instantiation, this can lead to unsatisfiable output buffer requirements.

This change makes it so that the block is not a `sync_interpolator` anymore, but simply keeps track of how many times the last item still has to be repeated.

On the way there:

- Changed block type from `sync_interpolator` to `block`
- Hence, implemented a correct `forecast`
- Constructor and interpolation setter check for interpolation >= 1
- `general_work`: aforementioned handling of partial repeats
- `qa_repeat`: original QA was totally insufficient and ignored most
  corner cases. Also, checking code was very slow.

Signed-off-by: Marcus Müller <mmueller@gnuradio.org>