runtime: Remove VLA macro, usage
VLA (Variable-Length Arrays) are a C (pre-C23) feature that only works
for some C++ compilers out of goodwill – in C++, an object's type is
known at compile time, and `int[3]` and `int[4]` are two different
types.

They do nothing special that std::vector couldn't do as well, and
there's only one usage within GNU Radio that should actually have used
allocation in the constructor, and a field. So, changed it that way.

Signed-off-by: Marcus Müller <mmueller@gnuradio.org>