blocks: fix repeat block tag propagation
The repeat block was refactored as a general block in #6813
(it was a sync_interpolator before). However, tag propagation was
missing from the refactor. Manual tag propagation needs to be done
if the block is a general block, because the automatic tag propagation
preserves the offset of tags, which is not what we want for the
repeat block.

Incidentally, before the refactor tag propagation was also somehow
broken for interpolation changes at runtime. After an interpolation change,
tags would disappear in the output for some time, or appear in the wrong
place.

The tag propagation behaviour introduced by this block is the most
natural one. Tags from each input item are copied to the first
repetition of such input item in the output.

Additionally, two QA tests for tag propagation have been added. One
of them does no interpolation changes at runtime, and the other
does one interpolation change at runtime. The two tests do not pass
with the previous version of the repeat block, but they pass after
adding this fix.

Signed-off-by: Daniel Estévez <daniel@destevez.net>
(cherry picked from commit cb5ce0add410b6cad239f9cd23675dc2010c1820)
Signed-off-by: Jeff Long <willcode4@gmail.com>