digital: add sps setter to symbol_sync
This adds a method to update the samples/symbol of the symbol_sync
block at runtime. When the set_sps() method is called, it resets
all the internal symbol synchronization loop status and ajusts values
for the new samples/symbol, so synchronization is lost and the loop
needs to lock again. This is reasonable, because the set_sps() method
can only be called asynchronously with respect to symbol rate changes
in the input sample stream. A more advanced runtime change could use
tags to update the samples/symbol at a specific point without losing
loop lock, but this commit does not cover this use case.

Signed-off-by: Daniel Estévez <daniel@destevez.net>
(cherry picked from commit 7303d30a3ffec2c84ea92f4529cc5d793050fff7)
Signed-off-by: Jeff Long <willcode4@gmail.com>